Refactor storage and improve UI
